Fix compression
Seems I was working on this in parallel to what got merged.

My reasoning is this: Use the expanded `-noX` commands that mean the same as `—no-compresion` according to the documentation.

By default use zstd compression instead of xz, because the AppImage team has said on at least one occassion that’s what’s going in the standard spec.

Signed-off-by: Gregorio Litenstein <g.litenstein@gmail.com>